+++ lib/src/date/show_date_picker_dialog.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../shared/picker_type.dart';
 import 'date_picker.dart';
 
 /// Shows a dialog containing a Material Design date picker.
