+++ lib/src/shared/picker_type.dart
@@ -1,0 +1,18 @@
+/// Initial display of the picker.
+///
+/// See also:
+///
+///  * [showRangeDatePickerDialog], which shows a dialog that contains a Material Design
+///    range picker.
+///  * [RangeDatePicker], widget which implements the Material Design Range picker.
+///
+enum PickerType {
+  /// Choosing a day.
+  days,
+
+  /// Choosing a month.
+  months,
+
+  /// Choosing a year.
+  years,
+}
