+++ lib/src/shared/year_picker.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'header.dart';
-import 'show_date_picker_dialog.dart';
+import '../date/show_date_picker_dialog.dart';
 import 'year_view.dart';
 
 /// A scrollable grid of years to allow picking a year.
