+++ test/year_picker_test.dart
@@ -1,5 +1,5 @@
-import 'package:date_picker_plus/src/header.dart';
-import 'package:date_picker_plus/src/year_picker.dart';
+import 'package:date_picker_plus/src/shared/header.dart';
+import 'package:date_picker_plus/src/shared/year_picker.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
